requirements.txt
streamlit==1.28.1
pandas>=2.2.0
plotly>=5.17.0
supabase==2.0.2
python-dotenv==1.0.0
reportlab>=4.0.0
pytz>=2021.1 